:art: Refactors picker showcase to use table view instead view controller
