+++ EmeraldComponentsApp/EmeraldComponentsApp/Commons/PickerData.swift
@@ -1,0 +1,16 @@
+//
+//  PickerData.swift
+//  EmeraldComponentsApp
+//
+//  Created by João Mendes | Stone on 11/09/18.
+//  Copyright © 2018 StoneCo. All rights reserved.
+//
+
+import EmeraldComponents
+
+
+extension String: PickerData {
+    public var identifier: String {
+        return self
+    }
+}
