+++ EmeraldComponentsApp/EmeraldComponentsApp/Components/Input/Input.storyboard
@@ -13,7 +13,7 @@
         <!--Input-->
         <scene sceneID="bBB-3j-2NE">
             <objects>
-                <tableViewController storyboardIdentifier="InputTableViewController" useStoryboardIdentifierAsRestorationIdentifier="YES" id="B8M-gB-blH" customClass="InputTableViewController" customModule="EmeraldComponentsApplication" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController storyboardIdentifier="InputTableViewController" useStoryboardIdentifierAsRestorationIdentifier="YES" id="B8M-gB-blH" customClass="InputTableViewController" customModule="EmeraldComponentsApp" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="static" style="grouped" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="18" sectionFooterHeight="18" id="loZ-Ns-TAn">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
