--- EmeraldComponentsApp/EmeraldComponentsApp/Components/PickerView/PickerView.storyboard
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14113" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="Srt-B4-R5S">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <capability name="Constraints with non-1.0 multipliers" minToolsVersion="5.1"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Picker-->
-        <scene sceneID="kMx-gY-4uW">
-            <objects>
-                <viewController storyboardIdentifier="PickerViewController" useStoryboardIdentifierAsRestorationIdentifier="YES" id="Srt-B4-R5S" customClass="PickerViewController" customModule="EmeraldComponentsApplication" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="DO3-dI-ZoL">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="LD6-8d-F3f" customClass="EmeraldPicker" customModule="EmeraldComponents">
-                                <rect key="frame" x="38" y="40" width="300" height="120"/>
-                                <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <userDefinedRuntimeAttributes>
-                                    <userDefinedRuntimeAttribute type="string" keyPath="title" value="Banco"/>
-                                    <userDefinedRuntimeAttribute type="string" keyPath="bottomText" value="Usado para depósitos de crédito"/>
-                                </userDefinedRuntimeAttributes>
-                            </view>
-                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="2W0-8r-NEF" customClass="EmeraldPicker" customModule="EmeraldComponents">
-                                <rect key="frame" x="38" y="180" width="300" height="51"/>
-                                <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <userDefinedRuntimeAttributes>
-                                    <userDefinedRuntimeAttribute type="string" keyPath="title" value="Banco Principal"/>
-                                </userDefinedRuntimeAttributes>
-                            </view>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="LD6-8d-F3f" firstAttribute="top" secondItem="33o-FZ-6aY" secondAttribute="top" constant="20" id="BC5-an-djf"/>
-                            <constraint firstItem="33o-FZ-6aY" firstAttribute="bottom" secondItem="2W0-8r-NEF" secondAttribute="bottom" constant="436" id="EHA-dl-nQF"/>
-                            <constraint firstItem="2W0-8r-NEF" firstAttribute="top" secondItem="LD6-8d-F3f" secondAttribute="bottom" constant="20" id="Mwf-hs-fWw"/>
-                            <constraint firstItem="LD6-8d-F3f" firstAttribute="centerX" secondItem="DO3-dI-ZoL" secondAttribute="centerX" id="a31-pd-dTj"/>
-                            <constraint firstItem="2W0-8r-NEF" firstAttribute="centerX" secondItem="DO3-dI-ZoL" secondAttribute="centerX" id="ea5-cI-fbV"/>
-                            <constraint firstItem="2W0-8r-NEF" firstAttribute="top" secondItem="DO3-dI-ZoL" secondAttribute="top" constant="180" id="efi-tZ-wp5"/>
-                            <constraint firstItem="LD6-8d-F3f" firstAttribute="width" secondItem="DO3-dI-ZoL" secondAttribute="width" multiplier="0.8" id="miE-BG-9ni"/>
-                            <constraint firstItem="2W0-8r-NEF" firstAttribute="width" secondItem="LD6-8d-F3f" secondAttribute="width" id="vgx-kx-BE5"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="33o-FZ-6aY"/>
-                    </view>
-                    <navigationItem key="navigationItem" title="Picker" largeTitleDisplayMode="never" id="gc3-Gi-s4z"/>
-                    <connections>
-                        <outlet property="firstPicker" destination="LD6-8d-F3f" id="6JJ-jh-JoA"/>
-                        <outlet property="secondPicker" destination="2W0-8r-NEF" id="Z3p-g2-fJ3"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="1kg-zj-0O8" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="1584" y="1183"/>
-        </scene>
-    </scenes>
-</document>
